requirements.txt
flask
flask-cors
openai
google-generativeai
gunicorn